Fix strict-types fallout in the networking stack; add URLSession test suite
Two coercions that weak mode hid became fatal TypeErrors with
declare(strict_types=1), both on the URLSession hot path:

- The CURL header callback fed CURLINFO_CONTENT_LENGTH_DOWNLOAD, which
  PHP reports as float (-1.0 while unknown, i.e. on the first header
  line of every response), straight into didReceiveHeaderData(int).
  Every request died on its first header; the value is now cast.
- Dictionary::firstIndex() returned the raw array key, but PHP
  converts numeric-string keys to int internally, so dictionaries
  keyed by number — TaskRegistry registers tasks by identifier —
  violated the declared ?string return when a transfer completed.
  The key is now cast to honor the contract.

tests/URLSessionTest.php exercises the stack end to end against PHP's
built-in server: body delivery and status 200 (the header regression),
JSON decoding, custom response headers, 404 delivered as a response
rather than a transport error, POST bodies, and connection refusal
surfacing as an Error. CollectionsTest gains regression checks for the
numeric-key Dictionary contract.

+++ src/Networking/EasyHandle.php
@@ -291,7 +291,7 @@ final class EasyHandle
             $this->updateProgressMeter(new EasyHandleProgress($totalBytesSent, $totalBytesExpectedToSend, $totalBytesReceived, $totalBytesExpectedToReceive));
             return CURLE_OK;
         }, CURLOPT_PROGRESSFUNCTION);
-        $this->set(fn(CurlHandle $handle, string $data): int => $this->didReceiveHeaderData($data, curl_getinfo($handle, CURLINFO_CONTENT_LENGTH_DOWNLOAD)), CURLOPT_HEADERFUNCTION);
+        $this->set(fn(CurlHandle $handle, string $data): int => $this->didReceiveHeaderData($data, (int)curl_getinfo($handle, CURLINFO_CONTENT_LENGTH_DOWNLOAD)), CURLOPT_HEADERFUNCTION);
     }
 
     public function urlErrorCode(int $easyCode): ?int
